% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzWrappers.R
\name{createJaspContainer}
\alias{createJaspContainer}
\alias{jaspContainerR}
\title{jaspContainer}
\usage{
createJaspContainer(
  title = "",
  dependencies = NULL,
  position = NULL,
  initCollapsed = FALSE
)
}
\arguments{
\item{title}{String, title of the jaspContainer}

\item{dependencies}{see jaspObject}

\item{position}{see jaspObject}

\item{initCollapsed}{Logical, should the container be collapsed by default?}
}
\description{
A jaspObject that can contain other jaspObjects
}
\examples{
library(jaspResults)

container <- createJaspContainer(title = "This is a container")
container$print()

container[["internal_name_1"]] <- createJaspContainer(title = "name shown to users 1")
container[["internal_name_2"]] <- createJaspContainer(title = "name shown to users 2")
container$print()

# let's reverse the position
container[["subcontainer1"]]$position <- 2
container[["subcontainer2"]]$position <- 1
container$print()
}
\section{Super classes}{
\code{\link[jaspResults:jaspObjR]{jaspResults::jaspObjR}} -> \code{\link[jaspResults:jaspOutputObjR]{jaspResults::jaspOutputObjR}} -> \code{jaspContainerR}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{initCollapsed}}{is the container collapsed?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{jaspContainerR$new()}}
\item \href{#method-length}{\code{jaspContainerR$length()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspObjR" data-id="dependOn">}\href{../../jaspResults/html/jaspObjR.html#method-dependOn}{\code{jaspResults::jaspObjR$dependOn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspObjR" data-id="print">}\href{../../jaspResults/html/jaspObjR.html#method-print}{\code{jaspResults::jaspObjR$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspOutputObjR" data-id="addCitation">}\href{../../jaspResults/html/jaspOutputObjR.html#method-addCitation}{\code{jaspResults::jaspOutputObjR$addCitation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspOutputObjR" data-id="getError">}\href{../../jaspResults/html/jaspOutputObjR.html#method-getError}{\code{jaspResults::jaspOutputObjR$getError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspOutputObjR" data-id="printHtml">}\href{../../jaspResults/html/jaspOutputObjR.html#method-printHtml}{\code{jaspResults::jaspOutputObjR$printHtml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspOutputObjR" data-id="setError">}\href{../../jaspResults/html/jaspOutputObjR.html#method-setError}{\code{jaspResults::jaspOutputObjR$setError()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new jaspContainer, although `createJaspContainer()` is preferred.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspContainerR$new(
  title = "",
  dependencies = NULL,
  position = NULL,
  info = NULL,
  initCollapsed = FALSE,
  jaspObject = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{String, title of the jaspContainer}

\item{\code{dependencies}}{see jaspObject}

\item{\code{position}}{see jaspObject}

\item{\code{info}}{string.}

\item{\code{initCollapsed}}{Logical, should the container be collapsed by default?}

\item{\code{jaspObject}}{Do not use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
The number of elements inside the container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspContainerR$length()}\if{html}{\out{</div>}}
}

}
}
