% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzWrappers.R
\name{jaspObjR}
\alias{jaspObjR}
\alias{jaspObject}
\alias{jaspObj}
\title{jaspObject}
\description{
The parent class of all jaspObjects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{jaspObjR$new()}}
\item \href{#method-print}{\code{jaspObjR$print()}}
\item \href{#method-dependOn}{\code{jaspObjR$dependOn()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Do not create an instance of a jaspObject (it will throw an error).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspObjR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print a jaspObject. Note that `print(jaspObj)` calls `jaspObj$print()`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspObjR$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dependOn"></a>}}
\if{latex}{\out{\hypertarget{method-dependOn}{}}}
\subsection{Method \code{dependOn()}}{
Specify dependencies for a jaspObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspObjR$dependOn(
  options = NULL,
  optionsFromObject = NULL,
  optionContainsValue = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{string, the name of an option, e.g., `"variables"`.}

\item{\code{optionsFromObject}}{jaspObject, copy the dependencies from another jaspObject.}

\item{\code{optionContainsValue}}{named list, the name corresponds to an option and the value corresponds to whatever this options should contain, e.g., `list(variables = "contNormal")`}
}
\if{html}{\out{</div>}}
}
}
}
