% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzWrappers.R
\name{jaspOutputObjR}
\alias{jaspOutputObjR}
\title{jaspOutputObjR}
\description{
Super object of jaspObjects that show output. Not to be used directly.
}
\section{Super class}{
\code{\link[jaspResults:jaspObjR]{jaspResults::jaspObjR}} -> \code{jaspOutputObjR}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{position}}{the position of a jaspObject, accepts numeric input}

\item{\code{title}}{the title of a jaspObject, accepts string input}

\item{\code{info}}{the info of a jaspObject, accepts string input}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{jaspOutputObjR$new()}}
\item \href{#method-printHtml}{\code{jaspOutputObjR$printHtml()}}
\item \href{#method-setError}{\code{jaspOutputObjR$setError()}}
\item \href{#method-getError}{\code{jaspOutputObjR$getError()}}
\item \href{#method-addCitation}{\code{jaspOutputObjR$addCitation()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspObjR" data-id="dependOn">}\href{../../jaspResults/html/jaspObjR.html#method-dependOn}{\code{jaspResults::jaspObjR$dependOn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="jaspResults" data-topic="jaspObjR" data-id="print">}\href{../../jaspResults/html/jaspObjR.html#method-print}{\code{jaspResults::jaspObjR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Do not use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspOutputObjR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-printHtml"></a>}}
\if{latex}{\out{\hypertarget{method-printHtml}{}}}
\subsection{Method \code{printHtml()}}{
print jaspObject as html.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspOutputObjR$printHtml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setError"></a>}}
\if{latex}{\out{\hypertarget{method-setError}{}}}
\subsection{Method \code{setError()}}{
set an error on this jaspObject. If this is a container then the message `x` will be shown on the first child that is a  plot or table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspOutputObjR$setError(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, }}{string, error message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getError"></a>}}
\if{latex}{\out{\hypertarget{method-getError}{}}}
\subsection{Method \code{getError()}}{
check whether a jaspObject has an error or not, returns a boolean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspOutputObjR$getError()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addCitation"></a>}}
\if{latex}{\out{\hypertarget{method-addCitation}{}}}
\subsection{Method \code{addCitation()}}{
set an citation this jaspObject. If this is a container then the citation `x` will also be set on all children.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jaspOutputObjR$addCitation(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, }}{string, citation.}
}
\if{html}{\out{</div>}}
}
}
}
