\name{jaspResults-package}
\alias{jaspResults-package}
\alias{jaspResults}
\docType{package}
\title{
  Easy results for your JASP analysis
}
\description{
  This package helps you build your analysis for JASP (http://jasp-stats.org).
  Using the included classes jaspResults, jaspTable and others you can easily structure output into tables, collections and plots.
  When you then later on run this in JASP it will automatically make it appear fully formatted in the resultsview. 
  When developing your analysis in R(-Studio) you can see a simpler representation by using $print().
}
\details{
  See the following pages for more information on how to use this package: \link{jaspObject}, \link{jaspContainer}, \link{jaspTable}, \link{jaspPlot}, \link{jaspHtml}, \link{jaspList} and of course \link{jaspResultsClass}.
}
\author{
  Joris Goosen j.c.goosen@uva.nl

Maintainer: JASP Team <info@jasp-stats.org>
}
